+++ R/strings.R
@@ -75,8 +75,8 @@ drake_strings <- function(...){
 #'   on a *file* (target or import), not an ordinary object.
 #' @description This function simply wraps literal double quotes around
 #'   the argument `x` so `drake` knows it is the name of a file.
-#'   Use when you are calling functions like `deps()`: for example,
-#'   `deps(file_store("report.md"))`. See the examples for details.
+#'   Use when you are calling functions like `deps_code()`: for example,
+#'   `deps_code(file_store("report.md"))`. See the examples for details.
 #'   Internally, `drake` wraps the names of file targets/imports
 #'   inside literal double quotes to avoid confusion between
 #'   files and generic R objects.
@@ -97,7 +97,7 @@ drake_strings <- function(...){
 #'   head(readd(small)) # You can use symbols for ordinary objects.
 #'   # But if you want to read cached info on files, use `file_store()`.
 #'   readd(file_store("report.md"), character_only = TRUE) # File fingerprint.
-#'   deps(file_store("report.Rmd"))
+#'   deps_code(file_store("report.Rmd"))
 #'   config <- drake_config(my_plan)
 #'   dependency_profile(file_store("report.Rmd"), config = config)
 #'   loadd(list = file_store("report.md"))